# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/projects": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "operationId": "GetProjectsAll",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.Project"
                            }
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/restapi.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "operationId": "PostProject",
                "parameters": [
                    {
                        "description": "Project Create Params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.ProjectCreateParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.CreatedItem"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/restapi.ResponseError"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "operationId": "PostUser",
                "parameters": [
                    {
                        "description": "User Create Params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.UserCreateParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.CreatedItem"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/restapi.ResponseError"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "operationId": "GetUserByID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.User"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/restapi.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "restapi.ResponseError": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                }
            }
        },
        "storage.CreatedItem": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                }
            }
        },
        "storage.Project": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "CreatedBy": {
                    "type": "integer"
                },
                "Description": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "Title": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                },
                "UpdatedBy": {
                    "type": "integer"
                }
            }
        },
        "storage.ProjectCreateParams": {
            "type": "object",
            "properties": {
                "CreatedBy": {
                    "type": "integer"
                },
                "Description": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                }
            }
        },
        "storage.User": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "Login": {
                    "type": "string"
                },
                "Password": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "storage.UserCreateParams": {
            "type": "object",
            "properties": {
                "Login": {
                    "type": "string"
                },
                "Password": {
                    "type": "string"
                }
            }
        }
    }
}